\name{plot.cluster}
\alias{plot.cluster}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot the Results of the post-GMYC tree}
\description{
lot the Results of the post-GMYC tree
}

\usage{
plot.cluster(tr, lthresh, show.tip.label = TRUE, show.node.label = FALSE, cex = 0.5)
}

\arguments{
  \item{tr}{A tree of ape tree class.}
  \item{lthresh}{CHECKA vector of threshold values}
  \item{show.tip.label}{ xxx }
  \item{show.node.label}{ xxx }
  \item{cex}{ xxx }
}

\details{
arguably
}

\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}

\author{Tomochika Fujisawa \email{tomochika.fujisawa05@imperial.ac.uk} and Tim Barraclough \email{t.barraclough@imperial.ac.uk}}

\seealso{\code{\link{gmyc}}}

\keyword{utilities}{plot.cluster}
